# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp/NFS-e API",
        "description": "API de integração entre agendamentos e providers WhatsApp/NFS-e.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor atual"
        }
    ],
    "paths": {
        "/api/v1/fiscal/accounts": {
            "get": {
                "tags": [
                    "Fiscal"
                ],
                "summary": "Lista credenciais fiscais por estabelecimento",
                "operationId": "c0fe486be946d673bc98381f3db506c7",
                "parameters": [
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de credenciais"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fiscal"
                ],
                "summary": "Cria/atualiza a credencial fiscal de um estabelecimento",
                "operationId": "193ae55314efdeba3a1629ca0c3bf54e",
                "parameters": [
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "establishment_external_id",
                                    "access_token"
                                ],
                                "properties": {
                                    "establishment_external_id": {
                                        "type": "string",
                                        "example": "est-001"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "example": "asaas"
                                    },
                                    "access_token": {
                                        "description": "Token da subconta (armazenado criptografado)",
                                        "type": "string"
                                    },
                                    "base_url": {
                                        "type": "string",
                                        "example": "https://api-sandbox.asaas.com",
                                        "nullable": true
                                    },
                                    "environment": {
                                        "type": "string",
                                        "example": "sandbox"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Credencial salva"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
        "/api/v1/inbound-messages": {
            "get": {
                "tags": [
                    "Inbound"
                ],
                "summary": "Lista mensagens recebidas (cliente → empresa)",
                "operationId": "2be2aba2d653d4eba0a0ae48a6936cc0",
                "parameters": [
                    {
                        "name": "wa_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "5519991170718"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mensagens recebidas"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
        "/api/v1/inbound-messages/{inbound}": {
            "get": {
                "tags": [
                    "Inbound"
                ],
                "summary": "Consulta uma mensagem recebida",
                "operationId": "abcf4409418fb9b5abd104cb0d542de9",
                "parameters": [
                    {
                        "name": "inbound",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem recebida"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "hmac": {
                "type": "apiKey",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}